Add GCF VM Delete Event Handler Example
This patch includes a solution developed in the field based on a
specific customer use case:  When a VM is deleted, DNS A records
associated with the VM should be automatically deleted.

The customer requested the solution be general enough to support VM
deletion through the Console UI, gcloud, and API.  Additionally, both
Deployment Manager and Terraform may used to manage the lifecycle of VM
resources.  The customer supports end-users creating resources with
Terraform or Deployment Manager and then manually deleting resources in
the Console, despite the recommendation to allow the IaC tooling to
manage the complete lifecycle..  In this situation, they want the DNS
records cleaned up automatically.

The solution includes a GCF Background Function handling
compute.instances.delete events.  The customer also required the
function compare the IP address of the VM's primary interface with the
DNS A record rrdatas.  This requirement introduces a race condition, the
function races against the VM delete operation.  In practice the
function wins the race most of the time and the cleanup is successful.